requirements.txt
info-nce-pytorch==0.1.4
matplotlib==3.10.3
numpy==1.26.4
pandas==2.3.1
peft==0.18.0
scikit-learn==1.7.1
scipy==1.15.3
seaborn==0.13.2
torch==2.7.1
tsaug==0.2.1
umap-learn==0.5.9.post2
yq==3.4.3
